eslint configure and fix
