+++ .eslintrc.json
@@ -11,6 +11,11 @@
         "sourceType": "module"
     },
     "rules": {
-        "no-console": "off"
+        "no-console": "off",
+        "no-restricted-syntax": [
+            "off",
+            "ForOfStatement"
+        ],
+        "no-await-in-loop": "off"
     }
 }
