+++ identifier.js
@@ -1,3 +1,3 @@
-export function getCommentPrefix(identifier) {
+export default function getCommentPrefix(identifier) {
   return `<!-- ${identifier}: do not delete/edit this line -->`;
 }
